Commit based on Sphere's review. - Fixed typo in documentation. - Normalized all function to have float instead of int as a parameter. - Normalized naming of functions to be 'Jump Height' related instead of 'Jump Force Level'. - Reduced RB_CalcMaximumJumpHeightTexCoords, but this will be changed next commit.

+++ code/renderer/tr_local.h
@@ -2206,6 +2206,6 @@ struct DDS_HEADER
 
 // Speed Academy
 void RE_SetPlayerJumpStartWorldZ(float value);
-void RE_SetPlayerJumpHeight(int value);
+void RE_SetPlayerJumpHeight(float value);
 
 #endif //TR_LOCAL_H
